# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  nickname: string;
}

export interface UserProfile {
  id: number;
  nickname: string;
  bubbleCount: number;
}

export interface LeaderboardEntry {
  rank: number;
  id: number;
  nickname: string;
  bubbleCount: number;
}

export interface LeaderboardResult {
  topTen: LeaderboardEntry[];
  myEntry?: LeaderboardEntry | null;
}

export type GetLeaderboardParams = {
userId?: number;
};