# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface LeaderboardEntry {
  rank: number;
  id: number;
  nickname: string;
  bubbleCount: number;
}