# lib/api-zod/src/generated/types/leaderboardResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { LeaderboardEntry } from './leaderboardEntry';

export interface LeaderboardResult {
  topTen: LeaderboardEntry[];
  myEntry?: LeaderboardEntry | null;
}