# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface UserInput {
  /**
     * @minLength 1
     * @maxLength 20
     */
  nickname: string;
}